requirements.txt
numpy
python-pptx
matplotlib
google-generativeai
PyMuPDF